# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_basilex_promenade_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Error'
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Pagination'
      status:
        type: string
    type: object
...
  internal_contexts_warehouse_location_adapter_http.CreateLocationRequest:
    properties:
      capacity:
        type: integer
      code:
        maxLength: 50
        type: string
      depth:
        type: number
      description:
        type: string
      height:
        type: number
      is_limited:
        type: boolean
      is_pickable:
        type: boolean
      is_putawayable:
        type: boolean
      name:
        maxLength: 200
        type: string
      notes:
        type: string
      parent_id:
        type: string
      type:
        enum:
        - warehouse
        - zone
        - aisle
        - rack
        - shelf
        - bin
        type: string
      width:
        type: number
    required:
    - code
    - name
    - type
    type: object
  internal_contexts_warehouse_location_adapter_http.LocationListResponse:
    properties:
      locations:
        items:
          $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  internal_contexts_warehouse_location_adapter_http.UpdateLocationRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 200
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: alexander.vasilenko@gmail.com
    name: API Support
    url: https://github.com/basilex/promenade
  description: Modern backend platform for customer management, orders, and business
    workflows with clean DDD architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Promenade Platform
  version: 0.1.0
paths:
...
  /warehouse/locations:
    get:
      consumes:
      - application/json
      description: Get paginated list of locations with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Location type filter
        in: query
        name: type
        type: string
      - description: Location status filter
        in: query
        name: status
        type: string
      - description: Available locations only
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: List locations
      tags:
      - Location
    post:
      consumes:
      - application/json
      description: Create a new warehouse location
      parameters:
      - description: Location creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.CreateLocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Create location
      tags:
      - Location
  /warehouse/locations/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Delete location
      tags:
      - Location
    get:
      consumes:
      - application/json
      description: Get location details by ID
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get location by ID
      tags:
      - Location
    put:
      consumes:
      - application/json
      description: Update location details
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.UpdateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Update location
      tags:
      - Location
  /warehouse/locations/{id}/activate:
    put:
      consumes:
      - application/json
      description: Set location status to active
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Activate location
      tags:
      - Location
  /warehouse/locations/{id}/capacity:
    put:
      consumes:
      - application/json
      description: Update location capacity and limited flag
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Capacity update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.UpdateCapacityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.CapacityResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Update capacity
      tags:
      - Location
  /warehouse/locations/{id}/children:
    get:
      consumes:
      - application/json
      description: Get direct or recursive child locations
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - default: false
        description: Include nested children
        in: query
        name: recursive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get location children
      tags:
      - Location
  /warehouse/locations/{id}/deactivate:
    put:
      consumes:
      - application/json
      description: Set location status to inactive
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Deactivate location
      tags:
      - Location
  /warehouse/locations/{id}/dimensions:
    put:
      consumes:
      - application/json
      description: Update physical dimensions of a location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Dimensions update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.UpdateDimensionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Update location dimensions
      tags:
      - Location
  /warehouse/locations/{id}/flags:
    put:
      consumes:
      - application/json
      description: Update operational flags for a location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Flags update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.UpdateFlagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_location_adapter_http.LocationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Update location flags
      tags:
      - Location
...
schemes:
- http
- https
swagger: "2.0"